Move optax_to_numpyro to optim
