+++ test/test_optimizers.py
@@ -10,6 +10,28 @@ import jax.numpy as jnp
 
 from numpyro import optim
 
+try:
+    import optax
+
+    # the optimizer test is parameterized by different optax optimizers, but we have
+    # to define them here to ensure that `optax` is defined. pytest.mark.parameterize
+    # decorators are run even if tests are skipped at the top of the file.
+    optax_optimizers = [
+        (optax.adam, (1e-2,), {}),
+        # clipped adam
+        (optax.chain, (optax.clip(10.0), optax.adam(1e-2)), {}),
+        (optax.adagrad, (1e-1,), {}),
+        # SGD with momentum
+        (optax.sgd, (1e-2,), {"momentum": 0.9}),
+        (optax.rmsprop, (1e-2,), {"decay": 0.95}),
+        # RMSProp with momentum
+        (optax.rmsprop, (1e-4,), {"decay": 0.9, "momentum": 0.9}),
+        (optax.sgd, (1e-2,), {}),
+    ]
+except ImportError:
+    pytestmark = pytest.mark.skip(reason="optax is not installed")
+    optax_optimizers = []
+
 
 def loss(params):
     return jnp.sum(params["x"] ** 2 + params["y"] ** 2)
@@ -23,20 +45,23 @@ def step(opt_state, optim):
 
 
 @pytest.mark.parametrize(
-    "optim_class, args",
+    "optim_class, args, kwargs",
     [
-        (optim.Adam, (1e-2,)),
-        (optim.ClippedAdam, (1e-2,)),
-        (optim.Adagrad, (1e-1,)),
-        (optim.Momentum, (1e-2, 0.5)),
-        (optim.RMSProp, (1e-2, 0.95)),
-        (optim.RMSPropMomentum, (1e-4,)),
-        (optim.SGD, (1e-2,)),
-    ],
+        (optim.Adam, (1e-2,), {}),
+        (optim.ClippedAdam, (1e-2,), {}),
+        (optim.Adagrad, (1e-1,), {}),
+        (optim.Momentum, (1e-2, 0.5), {}),
+        (optim.RMSProp, (1e-2, 0.95), {}),
+        (optim.RMSPropMomentum, (1e-4,), {}),
+        (optim.SGD, (1e-2,), {}),
+    ]
+    + optax_optimizers,
 )
-def test_optim_multi_params(optim_class, args):
+def test_optim_multi_params(optim_class, args, kwargs):
     params = {"x": jnp.array([1.0, 1.0, 1.0]), "y": jnp.array([-1, -1.0, -1.0])}
-    opt = optim_class(*args)
+    opt = optim_class(*args, **kwargs)
+    if not isinstance(opt, optim._NumPyroOptim):
+        opt = optim.optax_to_numpyro(opt)
     opt_state = opt.init(params)
     for i in range(2000):
         opt_state = step(opt_state, opt)
@@ -47,20 +72,23 @@ def test_optim_multi_params(optim_class, args):
 # note: this is somewhat of a bruteforce test. testing directly from
 # _NumpyroOptim would probably be better
 @pytest.mark.parametrize(
-    "optim_class, args",
+    "optim_class, args, kwargs",
     [
-        (optim.Adam, (1e-2,)),
-        (optim.ClippedAdam, (1e-2,)),
-        (optim.Adagrad, (1e-1,)),
-        (optim.Momentum, (1e-2, 0.5)),
-        (optim.RMSProp, (1e-2, 0.95)),
-        (optim.RMSPropMomentum, (1e-4,)),
-        (optim.SGD, (1e-2,)),
-    ],
+        (optim.Adam, (1e-2,), {}),
+        (optim.ClippedAdam, (1e-2,), {}),
+        (optim.Adagrad, (1e-1,), {}),
+        (optim.Momentum, (1e-2, 0.5), {}),
+        (optim.RMSProp, (1e-2, 0.95), {}),
+        (optim.RMSPropMomentum, (1e-4,), {}),
+        (optim.SGD, (1e-2,), {}),
+    ]
+    + optax_optimizers,
 )
-def test_numpyrooptim_no_double_jit(optim_class, args):
+def test_numpyrooptim_no_double_jit(optim_class, args, kwargs):
 
-    opt = optim_class(*args)
+    opt = optim_class(*args, **kwargs)
+    if not isinstance(opt, optim._NumPyroOptim):
+        opt = optim.optax_to_numpyro(opt)
     state = opt.init(jnp.zeros(10))
 
     my_fn_calls = 0
